+++ packages/gatsby/cache-dir/query-result-store.js
@@ -1,14 +1,13 @@
 import React from "react"
-
-import PageRenderer from "./page-renderer"
-import normalizePagePath from "./normalize-page-path"
 import { StaticQueryContext } from "gatsby"
 import {
-  getStaticQueryData,
   getPageQueryData,
   registerPath as socketRegisterPath,
   unregisterPath as socketUnregisterPath,
+  getStaticQueryData,
 } from "./socketIo"
+import PageRenderer from "./page-renderer"
+import normalizePagePath from "./normalize-page-path"
 
 if (process.env.NODE_ENV === `production`) {
   throw new Error(
@@ -26,19 +25,17 @@ const getPathFromProps = props =>
     ? normalizePagePath(props.pageResources.page.path)
     : undefined
 
-class JSONStore extends React.Component {
+export class PageQueryStore extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
-      staticQueryData: getStaticQueryData(),
       pageQueryData: getPageQueryData(),
       path: null,
     }
   }
 
-  handleMittEvent = (type, event) => {
+  handleMittEvent = () => {
     this.setState({
-      staticQueryData: getStaticQueryData(),
       pageQueryData: getPageQueryData(),
     })
   }
@@ -70,14 +67,12 @@ class JSONStore extends React.Component {
     // We want to update this component when:
     // - location changed
     // - page data for path changed
-    // - static query results changed
 
     return (
       this.props.location !== nextProps.location ||
       this.state.path !== nextState.path ||
       this.state.pageQueryData[normalizePagePath(nextState.path)] !==
-        nextState.pageQueryData[normalizePagePath(nextState.path)] ||
-      this.state.staticQueryData !== nextState.staticQueryData
+        nextState.pageQueryData[normalizePagePath(nextState.path)]
     )
   }
 
@@ -88,12 +83,44 @@ class JSONStore extends React.Component {
       return <div />
     }
 
+    return <PageRenderer {...this.props} {...data} />
+  }
+}
+
+export class StaticQueryStore extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      staticQueryData: getStaticQueryData(),
+    }
+  }
+
+  handleMittEvent = () => {
+    this.setState({
+      staticQueryData: getStaticQueryData(),
+    })
+  }
+
+  componentDidMount() {
+    ___emitter.on(`*`, this.handleMittEvent)
+  }
+
+  componentWillUnmount() {
+    ___emitter.off(`*`, this.handleMittEvent)
+  }
+
+  shouldComponentUpdate(nextProps, nextState) {
+    // We want to update this component when:
+    // - static query results changed
+
+    return this.state.staticQueryData !== nextState.staticQueryData
+  }
+
+  render() {
     return (
       <StaticQueryContext.Provider value={this.state.staticQueryData}>
-        <PageRenderer {...this.props} {...data} />
+        {this.props.children}
       </StaticQueryContext.Provider>
     )
   }
 }
-
-export default JSONStore